+++ spec/filters/agilibox/small_data/filter_strategy_by_date_or_datetime_period_spec.rb
@@ -1,0 +1,7 @@
+require "rails_helper"
+
+describe Agilibox::SmallData::FilterStrategyByDateOrDatetimePeriod do
+  it "should raise error" do
+    expect { described_class.new(:date_field) }.to raise_error(RuntimeError, /please use/)
+  end
+end