+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next version
 
 - Fix flash partial
+- Filter by date/datetime period now allows custom periods
 
 ## 1.2.0